Minor change to credential request
Early advice was for the Hub to clearly ID itself in the integrations that it supports, i.e.,  with an Application Name like "NZ ORCID Hub for Royal Society".  Since then the practice has changed to being a little less cluttered.  (Unless we instruct them otherwise) the ORCID team will just trim the app_name to being the organisation.name anyway so let's not send anything else.