Fix iD in header
ORCID IDs -> ORCID iDs
What I couldn't find the fix for as I think I understand the behaviour, just not the solution: When user first hits the linking page (/link), it's not a item in the header so nothing is greyed.  Should "your ORCID hub connections" be greyed then as it is when /profile is the location?