Fix Typos/Style in Login, About, and FAQ
A bunch of typo were introduced with the new user log in flow and
the addition of the FAQ page.  These fixed, and some styling added
to control the most ugly of the text not wrapping.

TO DO: take another look at what should be put this style into the
css, and specifying  the various col-XX-# so the site's a litle
more mobile friendly.