URL-encode PATH_INFO for /auth/<path:url>
This change requires the addition of AllowEncodedSlashes NoDecode
in Apache or its equivelent in other web servers.

However, without it Apache will automatically rewrite the double-slash
in the scheme (https://) into a single slash (http:/).  This in
turn causes flask to see the URL encoded as PATH_INFO in the
/auth/<path> route as a relative path rather than an absolute URL.